Add files via upload
Added inbox.ui file. Next step is to link the login.ui file to the inbox.ui file. To do this we will need to edit the main.py file and add a .clicked function on the login button so that when it is clicked, the inbox.ui file displays and the login.ui file is hidden.